+++ network/src/iosMain/kotlin/com/paydock/core/network/NetworkClientBuilder.ios.kt
@@ -2,13 +2,15 @@
 package com.paydock.core.network
 
 import com.paydock.core.network.dto.error.toApiError
-import com.paydock.core.network.exceptions.UnknownApiException
+import com.paydock.core.network.exceptions.ApiException
+import com.paydock.core.network.exceptions.ApiParseException
 import com.paydock.core.network.extensions.convertToApiErrorResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.engine.darwin.Darwin
 import io.ktor.client.engine.darwin.certificates.CertificatePinner
+import io.ktor.client.plugins.HttpRequestRetry
 import io.ktor.client.plugins.HttpResponseValidator
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.logging.DEFAULT
@@ -31,8 +33,18 @@ internal class IOSNetworkClientBuilder : NetworkClientBuilder() {
      * @throws IllegalArgumentException if the base URL is not set.
      */
     override fun build(): HttpClient {
-        requireNotNull(baseUrl) { "Base URL must be set" }
-        val config = NetworkConfig(baseUrl!!, sslPins, isDebug, protocol, requestTimeout, responseTimeout)
+        require(!baseUrl.isNullOrBlank()) { "Base URL must be set" }
+        val config =
+            NetworkConfig(
+                baseUrl!!,
+                sslPins,
+                protocol,
+                requestTimeout,
+                responseTimeout,
+                maxRetries,
+                retryInterval,
+                isDebug
+            )
 
         val engine = mockEngine ?: createHttpEngine(config)
         return HttpClient(engine) {
@@ -53,10 +65,26 @@ internal class IOSNetworkClientBuilder : NetworkClientBuilder() {
                         // Process the error data class accordingly and throw the appropriate ApiException
                         throw when {
                             apiErrorResponse != null -> apiErrorResponse.toApiError()
-                            else -> UnknownApiException(status = response.status.value, errorBody = errorBody)
+                            else -> ApiParseException(status = response.status.value, errorBody = errorBody)
                         }
                     } else {
-                        throw UnknownApiException(status = response.status.value, errorBody = errorBody)
+                        throw ApiParseException(status = response.status.value, errorBody = errorBody)
+                    }
+                }
+            }
+            // Only add retry logic if maxRetries is greater than 0
+            if (config.maxRetries > 0) {
+                install(HttpRequestRetry) {
+                    maxRetries = config.maxRetries
+                    retryOnServerErrors(maxRetries = config.maxRetries)
+                    retryIf { _, response -> !response.status.isSuccess() }
+                    retryOnExceptionIf { _, cause -> cause is ApiException }
+                    delayMillis { retry -> retry * config.retryInterval }
+                    modifyRequest { request ->
+                        request.headers.append(
+                            "x-retry-count",
+                            retryCount.toString()
+                        )
                     }
                 }
             }